+++ astropy/io/misc/asdf/tags/transform/compound.py
@@ -14,8 +14,6 @@ from ......modeling.models import Identity, Mapping
 
 __all__ = ['CompoundType', 'RemapAxesType']
 
-
-
 _operator_to_tag_mapping = {
     '+'  : 'add',
     '-'  : 'subtract',
@@ -23,7 +21,8 @@ _operator_to_tag_mapping = {
     '/'  : 'divide',
     '**' : 'power',
     '|'  : 'compose',
-    '&'  : 'concatenate'
+    '&'  : 'concatenate',
+    'set_inputs': 'set_inputs',
 }
 
 
@@ -34,7 +33,8 @@ _tag_to_method_mapping = {
     'divide'      : '__truediv__',
     'power'       : '__pow__',
     'compose'     : '__or__',
-    'concatenate' : '__and__'
+    'concatenate' : '__and__',
+    'set_inputs'  : 'set_inputs',
 }
 
 
@@ -55,10 +55,13 @@ class CompoundType(TransformType):
                 node['forward'][0]._tag))
         right = yamlutil.tagged_tree_to_custom_tree(
             node['forward'][1], ctx)
-        if not (isinstance(right, Model)):
+        if not (isinstance(right, Model)) or not (isinstance(right, dict)):
             raise TypeError("Unknown model type '{0}'".format(
                 node['forward'][1]._tag))
-        model = getattr(left, oper)(right)
+        if oper == 'set_inputs' and isinstance(right, dict):
+            model = CompoundModel('set_inputs', left, right)
+        else
+            model = getattr(left, oper)(right)
 
         model = cls._from_tree_base_transform_members(model, node, ctx)
         model.map_parameters()