+++ astropy/modeling/core.py
@@ -2348,7 +2348,7 @@ class CompoundModel(Model):
             self._user_inverse = inverse
         else:
             self._user_inverse = None
-        if op != '%' and len(left) != len(right):
+        if op != 'set_inputs' and len(left) != len(right):
             raise ValueError(
                 'Both operands must have equal values for n_models')
         else:
@@ -2389,51 +2389,52 @@ class CompoundModel(Model):
                                               self.right.inverse,
                                               self.left.inverse,
                                               inverse=self)
-        ###elif op == '%':
-        ###    if not isinstance(right, dict):
-        ###        raise ValueError('expecting dictionary for right side of "%"'
-        ###                         ' operator')
-            # else:
-            #     # Dict keys must match either possible indices
-            #     # for model on left side,
-            #     # or names for inputs.
-            #     self.n_inputs = left.n_inputs - len(right)
-            #     self.outputs = left.outputs
-            #     self.n_outputs = left.n_outputs
-            #     newinputs = list(left.inputs)
-            #     keys = right.keys()
-            #     input_ind = []
-            #     for key in keys:
-            #         if isinstance(key, int):
-            #             if key >= left.n_inputs or key < 0:
-            #                 raise ValueError(
-            #                     'substitution key integer value '
-            #                     'not among possible input choices')
-            #             else:
-            #                 if key in input_ind:
-            #                     raise ValueError("Duplicate specification of "
-            #                                      "same input (index/name)")
-            #                 else:
-            #                     input_ind.append(key)
-            #         elif isinstance(key, str):
-            #             if key not in left.inputs:
-            #                 raise ValueError(
-            #                     'Substitution key string not among possible '
-            #                     'input choices')
-            #             # Check to see it doesn't match positional
-            #             # specification.
-            #             ind = left.inputs.index(key)
-            #             if ind in input_ind:
-            #                 raise ValueError("Duplicate specification of "
-            #                                  "same input (index/name)")
-            #             else:
-            #                 input_ind.append(ind)
-            #     # Remove substituted inputs
-            #     input_ind.sort()
-            #     input_ind.reverse()
-            #     for ind in input_ind:
-            #         del newinputs[ind]
-            #     self.inputs = tuple(newinputs)
+        elif op == 'set_inputs':
+            if not isinstance(left, Model):
+                raise ValueError('First argument must be an instance of an astropy Model')
+            if not isinstance(right, dict):
+                raise ValueError('expecting dictionary for second argument')
+
+            # Dict keys must match either possible indices
+            # for model on left side,
+            # or names for inputs.
+            self.n_inputs = left.n_inputs - len(right)
+            self.outputs = left.outputs
+            self.n_outputs = left.n_outputs
+            newinputs = list(left.inputs)
+            keys = right.keys()
+            input_ind = []
+            for key in keys:
+                if isinstance(key, int):
+                    if key >= left.n_inputs or key < 0:
+                        raise ValueError(
+                            'substitution key integer value '
+                            'not among possible input choices')
+                    else:
+                        if key in input_ind:
+                            raise ValueError("Duplicate specification of "
+                                             "same input (index/name)")
+                        else:
+                            input_ind.append(key)
+                elif isinstance(key, str):
+                    if key not in left.inputs:
+                        raise ValueError(
+                            'Substitution key string not among possible '
+                            'input choices')
+                    # Check to see it doesn't match positional
+                    # specification.
+                    ind = left.inputs.index(key)
+                    if ind in input_ind:
+                        raise ValueError("Duplicate specification of "
+                                         "same input (index/name)")
+                    else:
+                        input_ind.append(ind)
+            # Remove substituted inputs
+            input_ind.sort()
+            input_ind.reverse()
+            for ind in input_ind:
+                del newinputs[ind]
+            self.inputs = tuple(newinputs)
 
         else:
             raise ModelDefinitionError('Illegal operator: ', self.op)
@@ -2497,7 +2498,7 @@ class CompoundModel(Model):
 
     def __call__(self, *args, **kw):
         op = self.op
-        if op != '%':
+        if op != 'set_inputs':
             if op != '&':
                 leftval = self.left(*args, **kw)
                 if op != '|':
@@ -2528,46 +2529,47 @@ class CompoundModel(Model):
                     return self.right(leftval, **kw)
             elif op in SPECIAL_OPERATORS:
                 return binary_operation(SPECIAL_OPERATORS[op], leftval, rightval)
-        # elif op == '%':
-        #     subs = self.right
-        #     newargs = list(args)
-        #     subinds = []
-        #     subvals = []
-        #     for key in subs.keys():
-        #         if isinstance(key, int):
-        #             subinds.append(key)
-        #         elif isinstance(key, str):
-        #             ind = self.left.inputs.index(key)
-        #             subinds.append(ind)
-        #         subvals.append(subs[key])
-        #     # Turn inputs specified in kw into positional indices.
-        #     # Names for compound inputs do not propagate to sub models.
-        #     kwind = []
-        #     kwval = []
-        #     for kwkey in list(kw.keys()):
-        #         if kwkey in self.inputs:
-        #             ind = self.inputs.index(kwkey)
-        #             if ind < len(args):
-        #                 raise ValueError("Keyword argument duplicates "
-        #                                 "positional value supplied")
-        #             kwind.append(ind)
-        #             kwval.append(kw[kwkey])
-        #             del kw[kwkey]
-        #     # Build new argument list
-        #     # Append keyword specified args first
-        #     if kwind:
-        #         kwargs = list(zip(kwind, kwval))
-        #         kwargs.sort()
-        #         kwindsorted, kwvalsorted = list(zip(*kwargs))
-        #         newargs = newargs + list(kwvalsorted)
-        #     if subinds:
-        #         subargs = list(zip(subinds, subvals))
-        #         subargs.sort()
-        #         subindsorted, subvalsorted = list(zip(*subargs))
-        #     # The substitutions must be inserted in order
-        #     for ind, val in subargs:
-        #         newargs.insert(ind, val)
-        #     return self.left(*newargs, **kw)
+        elif op == 'set_inputs':
+            subs = self.right
+            newargs = list(args)
+            subinds = []
+            subvals = []
+            for key in subs.keys():
+                if isinstance(key, int):
+                    subinds.append(key)
+                elif isinstance(key, str):
+                    ind = self.left.inputs.index(key)
+                    subinds.append(ind)
+                subvals.append(subs[key])
+            # Turn inputs specified in kw into positional indices.
+            # Names for compound inputs do not propagate to sub models.
+            kwind = []
+            kwval = []
+            for kwkey in list(kw.keys()):
+                if kwkey in self.inputs:
+                    ind = self.inputs.index(kwkey)
+                    if ind < len(args):
+                        raise ValueError("Keyword argument duplicates "
+                                        "positional value supplied")
+                    kwind.append(ind)
+                    kwval.append(kw[kwkey])
+                    del kw[kwkey]
+            # Build new argument list
+            # Append keyword specified args first
+            if kwind:
+                kwargs = list(zip(kwind, kwval))
+                kwargs.sort()
+                kwindsorted, kwvalsorted = list(zip(*kwargs))
+                newargs = newargs + list(kwvalsorted)
+            if subinds:
+                subargs = list(zip(subinds, subvals))
+                subargs.sort()
+                subindsorted, subvalsorted = list(zip(*subargs))
+            # The substitutions must be inserted in order
+            for ind, val in subargs:
+                newargs.insert(ind, val)
+            return self.left(*newargs, **kw)
+
         else:
             raise ModelDefinitionError('unrecognized operator')
 
@@ -3295,6 +3297,23 @@ try:
     import asdf_compound
 except ImportError:
     pass
+
+
+def set_inputs(modelinstance, values):
+    """
+    This function creates a compound model with one or more of the input values
+    of the input model assigned fixed values (scalar or array).
+
+    The values argument is expected to be a dictionary. If the key is an integer
+    it refers to the number of the input argument (0-based). If the key is a
+    string, it looks for an input of that name.
+
+    For example, if one defines g as Gaussian2D(1,2,3,4,5), then:
+
+    gv = set_inputs(g, {0: 2.5}) results in a 1D function equivalent to
+    Gaussian2D(1, 2, 3, 4, 5)(x=2.5, y)
+    """
+    return CompoundModel('set_inputs', modelinstance, values)
 
 
 def custom_model(*args, fit_deriv=None, **kwargs):